requirements.txt
Django>=3.1
Pillow
psycopg2